added everything and much more.
I'm sure we didn't need all the pictures, but I had them all in my lib
folder already.  That's why the file size is so huge!  I put them on
there, intentionally my idea was to edit all the pictures one by one so
that they could all be used in a portfolio section for each individual
tatoo artist.